GPA works with float, decimal, now. Still need to get the float to stop at two decimal places.

+++ db/migrate/20170823144457_add_gpa_to_students.rb
@@ -1,5 +1,5 @@
 class AddGpaToStudents < ActiveRecord::Migration[5.1]
   def change
-    add_column :students, :gpa, :integer
+    add_column :students, :gpa, :float
   end
 end
